refactor: moved fns to utils & modified bitcoin api config
